Rework handling templates
Do not hardcode available templates, build the list dynamically instead.
The main motivation is supporting various template flavors, where the
name is no longer just "template-version", but can be also
"template-version-flavor". But this change makes it also possible to
build different ISO image with different templates set included and they
will show up in the initial setup UI.

For kickstart settings, cut-out the template version, so it's possible
to write "fedora-xfce" and it will install the version that is available
on the installation image. The version-less name representation is used
internally as the primary template identifier.

Keep handling Whonix specially, because it's two templates and also it
isn't suitable for default template (for example it cannot function as
template for sys-net).

QubesOS/qubes-issues#7784